refactor: move keys back to walrus-core
