+++ crates/walrus-core/src/lib.rs
@@ -47,6 +47,7 @@ use thiserror::Error;
 pub mod bft;
 pub mod encoding;
 pub mod inconsistency;
+pub mod keys;
 pub mod merkle;
 pub mod messages;
 pub mod metadata;
