+++ crates/walrus-service/Cargo.toml
@@ -66,7 +66,6 @@ path = "bin/aggregator.rs"
 [dev-dependencies]
 mockall = "0.12.1"
 rand.workspace = true
-serde_test = "1.0.176"
 tempfile.workspace = true
 walrus-core = { workspace = true, features = ["test-utils"] }
 walrus-test-utils.workspace = true
